Combine JWT Server Filter with AuthenticationFilter
