+++ src/main/java/io/antmedia/console/rest/AuthenticationFilter.java
@@ -1,6 +1,7 @@
 package io.antmedia.console.rest;
 
 import java.io.IOException;
+import java.security.interfaces.RSAPublicKey;
 
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
@@ -11,6 +12,16 @@ import javax.servlet.http.HttpServletResponse;
 import javax.ws.rs.HttpMethod;
 
 import org.springframework.web.context.ConfigurableWebApplicationContext;
+
+import com.auth0.jwk.Jwk;
+import com.auth0.jwk.JwkException;
+import com.auth0.jwk.JwkProvider;
+import com.auth0.jwk.UrlJwkProvider;
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.JWTVerifier;
+import com.auth0.jwt.algorithms.Algorithm;
+import com.auth0.jwt.exceptions.JWTVerificationException;
+import com.auth0.jwt.interfaces.DecodedJWT;
 
 import io.antmedia.console.datastore.ConsoleDataStoreFactory;
 import io.antmedia.console.datastore.AbstractConsoleDataStore;
@@ -23,6 +34,7 @@ import io.antmedia.settings.ServerSettings;
 public class AuthenticationFilter extends AbstractFilter {
 
 	public static final String DISPATCH_PATH_URL = "_path";
+	public static final String JWT_TOKEN = "Authorization";
 
 	public AbstractConsoleDataStore getDataStore()
 	{
@@ -85,10 +97,22 @@ public class AuthenticationFilter extends AbstractFilter {
 
 		ServerSettings serverSettings = getServerSetting();
 		
-		// If it's passed from JWT Token then bypass Authentication Filter
+		/*
+		 * JWT Filter checking below parameters:
+		 * JWT Server Filter enable or not 
+		 * Is Token filled or not
+		 * Is JWT Server Token valid or invalid 
+		 */
 		if (serverSettings != null && serverSettings.isJwtServerControlEnabled()
-				&& (httpRequest.getHeader(JWTServerFilter.JWT_TOKEN) != null)) {
+				&& (httpRequest.getHeader(JWT_TOKEN) != null)
+				&& checkJWT(httpRequest.getHeader(JWT_TOKEN))
+				) {
 			chain.doFilter(request, response);
+		}
+		else if(serverSettings != null && serverSettings.isJwtServerControlEnabled()
+				&& (httpRequest.getHeader(JWT_TOKEN) != null)
+				&& !checkJWT(httpRequest.getHeader(JWT_TOKEN))) {
+			((HttpServletResponse) response).sendError(HttpServletResponse.SC_FORBIDDEN, "Invalid JWT Server Token");
 		}
 		else if (path.equals("/rest/isAuthenticated") ||
 				path.equals("/rest/authenticateUser") || 
@@ -212,4 +236,36 @@ public class AuthenticationFilter extends AbstractFilter {
 		}
 		return granted;
 	}
+	
+
+	private boolean checkJWT( String jwtString) {
+		boolean result = true;
+		try {
+			String jwksURL = getServerSetting().getJwksURL();
+
+			if (jwksURL != null && !jwksURL.isEmpty()) {
+				DecodedJWT jwt = JWT.decode(jwtString);
+				JwkProvider provider = new UrlJwkProvider(jwksURL);
+				Jwk jwk = provider.get(jwt.getKeyId());
+				Algorithm algorithm = Algorithm.RSA256((RSAPublicKey) jwk.getPublicKey(), null);
+				algorithm.verify(jwt);
+			}
+			else {
+				Algorithm algorithm = Algorithm.HMAC256(getServerSetting().getJwtServerSecretKey());
+				JWTVerifier verifier = JWT.require(algorithm)
+						.build();
+				verifier.verify(jwtString);
+			}
+
+		}
+		catch (JWTVerificationException ex) {
+			logger.error(ex.toString());
+			result = false;
+		} catch (JwkException e) {
+			logger.error(e.toString());
+			result = false;
+		}
+		return result;
+	}
+	
 }
