+++ src/test/java/io/antmedia/test/console/AuthenticationFilterTest.java
@@ -21,6 +21,9 @@ import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
 import org.springframework.web.context.ConfigurableWebApplicationContext;
 
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.algorithms.Algorithm;
+
 import io.antmedia.console.datastore.AbstractConsoleDataStore;
 import io.antmedia.console.rest.AuthenticationFilter;
 import io.antmedia.console.rest.CommonRestService;
@@ -37,7 +40,7 @@ public class AuthenticationFilterTest {
 		
 		ServerSettings serverSettings = new ServerSettings();
 		Mockito.doReturn(serverSettings).when(filter).getServerSetting();
-
+		
 		HttpServletRequest request = Mockito.mock(HttpServletRequest.class);
 		HttpServletResponse response =  Mockito.mock(HttpServletResponse.class);
 		FilterChain chain = Mockito.mock(FilterChain.class);
@@ -284,33 +287,66 @@ public class AuthenticationFilterTest {
 	    	
 	    	AuthenticationFilter authenticationFilter = Mockito.spy(new AuthenticationFilter());
 	        
-	        MockHttpServletResponse httpServletResponse;
+	    	HttpServletResponse httpServletResponse;
 	        MockHttpServletRequest httpServletRequest;
 	        MockFilterChain filterChain;
 	        
 	        ServerSettings serverSettings = new ServerSettings();
 	        serverSettings.setJwtServerControlEnabled(true);
+	        serverSettings.setJwtServerSecretKey("testtesttesttesttesttesttesttest");
 	        Mockito.doReturn(serverSettings).when(authenticationFilter).getServerSetting();
+	        
+	        String validToken = JWT.create().sign(Algorithm.HMAC256(serverSettings.getJwtServerSecretKey()));
+	        String invalidToken = JWT.create().sign(Algorithm.HMAC256("invalid-key-invalid-key-invalid-key"));
+	        String appJwtToken = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.e30._dEfIoSDLBJo49hmyivqYziMIGeWOD19Ex0kGGK_6ww";
+	        String jwkstoken = "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6InFWMWhqT0o2RFlFOWIzRDFSU0lHSSJ9.eyJpc3MiOiJodHRwczovL2FudG1lZGlhLnVzLmF1dGgwLmNvbS8iLCJzdWIiOiI3UVEzWTlLSzJPY1dOSzRwMXpydGU1UzQxSWR4amxLc0BjbGllbnRzIiwiYXVkIjoiaHR0cHM6Ly9hbnRtZWRpYS51cy5hdXRoMC5jb20vYXBpL3YyLyIsImlhdCI6MTYyODE1MTQ4NSwiZXhwIjoxNjI4MjM3ODg1LCJhenAiOiI3UVEzWTlLSzJPY1dOSzRwMXpydGU1UzQxSWR4amxLcyIsImd0eSI6ImNsaWVudC1jcmVkZW50aWFscyJ9.dgtT0dqL_JiA1AJRIWYyJMU7KG_EpJzucqlmIdEt36rL35G9QLWcxJVWCM-OFDAje9UaNDqFVHMNfXzDhvXrs5LvPlEFSZVAZUtMgjP0X94hlCjKIrvhnfN2lcDZFsSMIqXJeoPMjlRvGItrphRQaMr5ow3eCcvRYVK1MXvptGisdh1rTVBWPRRvaFH8x5yISw98DwNauVWW949o8JDIkortDQEXHmpipC7NoACCzZmGlmBv6ubaZxyTwh7QAp68kx6_tIcmj7nm6cLdoheFjH-io-ee6oTkLRr2krRqSjJrY9A_hJYH4Gixpe-F7mMeE8dMuRGhWue3pfMJmy8ulQ";
 
    			AbstractConsoleDataStore store = Mockito.mock(AbstractConsoleDataStore.class);
    			Mockito.doReturn(store).when(authenticationFilter).getDataStore();
 
 	        // JWT Token null && JWT Server filter enable scenario
+   			// It will try to login with user/pass details
 	        {   
 	        	//reset filterchain
 	        	filterChain = new MockFilterChain();
 	        	
 	        	//reset httpServletResponse
-	        	httpServletResponse = new MockHttpServletResponse();
+	        	httpServletResponse = Mockito.spy(new MockHttpServletResponse());
 	        	
 	        	//reset httpServletRequest
 	        	httpServletRequest = new MockHttpServletRequest();
 	        	
 	            authenticationFilter.doFilter(httpServletRequest, httpServletResponse, filterChain);
 	            assertEquals(HttpStatus.FORBIDDEN.value(),httpServletResponse.getStatus());
+				Mockito.verify(httpServletResponse).sendError(HttpServletResponse.SC_FORBIDDEN, "Not authenticated user");
 	        }
 	        
-	        // JWT Token filled && JWT Server filter enable scenario
+	        // JWT Token disable and valid token scenario
+	        {
+	            //reset filterchains
+	            filterChain = new MockFilterChain();
+
+	            //reset httpServletResponses
+	        	httpServletResponse = Mockito.spy(new MockHttpServletResponse());
+
+	            //reset httpServletRequest
+	            httpServletRequest = new MockHttpServletRequest();
+
+	            serverSettings.setJwtServerControlEnabled(false);
+
+	            Mockito.doReturn(serverSettings).when(authenticationFilter).getServerSetting();
+
+	            httpServletRequest.addHeader("Authorization", validToken);
+
+	            authenticationFilter.doFilter(httpServletRequest, httpServletResponse, filterChain);
+	            assertEquals(HttpStatus.FORBIDDEN.value(),httpServletResponse.getStatus());
+	            Mockito.verify(httpServletResponse).sendError(HttpServletResponse.SC_FORBIDDEN, "Not authenticated user");
+	        }
+	        
+	        //Restore JWT Server Control to true
+        	serverSettings.setJwtServerControlEnabled(true);
+	        
+	        // JWT Token filled && JWT Server filter enable && invalid token scenario
 	        {   
 	        	//reset filterchain
 	        	filterChain = new MockFilterChain();
@@ -321,7 +357,24 @@ public class AuthenticationFilterTest {
 	        	//reset httpServletRequest
 	        	httpServletRequest = new MockHttpServletRequest();
 	        	
-	            httpServletRequest.addHeader("Authorization", "testtest");
+	            httpServletRequest.addHeader("Authorization", invalidToken);
+
+	            authenticationFilter.doFilter(httpServletRequest, httpServletResponse, filterChain);
+	            assertEquals(HttpStatus.FORBIDDEN.value(),httpServletResponse.getStatus());
+	        }
+	        
+	        // JWT Token filled && JWT Server filter enable && valid token scenario
+	        {   
+	        	//reset filterchain
+	        	filterChain = new MockFilterChain();
+	        	
+	        	//reset httpServletResponse
+	        	httpServletResponse = new MockHttpServletResponse();
+	        	
+	        	//reset httpServletRequest
+	        	httpServletRequest = new MockHttpServletRequest();
+	        	
+	            httpServletRequest.addHeader("Authorization", validToken);
 
 	            authenticationFilter.doFilter(httpServletRequest, httpServletResponse, filterChain);
 	            assertEquals(HttpStatus.OK.value(),httpServletResponse.getStatus());
@@ -339,12 +392,56 @@ public class AuthenticationFilterTest {
 	        	httpServletRequest = new MockHttpServletRequest();
 	        	
 		        httpServletRequest.setRequestURI("/rest/v2/authentication-status");
-	        	
-	            serverSettings.setJwtServerControlEnabled(true);
-	            Mockito.doReturn(serverSettings).when(authenticationFilter).getServerSetting();
 	            
 	            authenticationFilter.doFilter(httpServletRequest, httpServletResponse, filterChain);
 	            assertEquals(HttpStatus.OK.value(),httpServletResponse.getStatus());
 	        }  
+	        
+	        // Valid JWKS Token Scenario
+	        {
+	            //reset filterchains
+	            filterChain = new MockFilterChain();
+
+	            //reset httpServletResponses
+	            httpServletResponse = new MockHttpServletResponse();
+
+	            //reset httpServletRequest
+	            httpServletRequest = new MockHttpServletRequest();
+
+	            httpServletRequest.setRemoteAddr("11.11.11.11");
+	            serverSettings.setJwksURL("https://antmedia.us.auth0.com");
+	            serverSettings.setJwtServerSecretKey("4Hr7PWwrTf6YFynkO5QeNQrlxe5r7HtfUdLhis2i_vbXdtF1VI0SwnP0ZSlhf0Yh");
+
+	            Mockito.doReturn(serverSettings).when(authenticationFilter).getServerSetting();
+
+	            httpServletRequest.addHeader("Authorization", jwkstoken);
+
+	            authenticationFilter.doFilter(httpServletRequest, httpServletResponse, filterChain);
+	            assertEquals(HttpStatus.OK.value(),httpServletResponse.getStatus());
+	        }
+	        
+	        // JWKS null scenario
+	        {
+	            //reset filterchains
+	            filterChain = new MockFilterChain();
+
+	            //reset httpServletResponses
+	            httpServletResponse = new MockHttpServletResponse();
+
+	            //reset httpServletRequest
+	            httpServletRequest = new MockHttpServletRequest();
+
+	            httpServletRequest.setRemoteAddr("11.11.11.11");
+	            serverSettings.setJwksURL("");
+
+	            Mockito.doReturn(serverSettings).when(authenticationFilter).getServerSetting();
+	            serverSettings.setJwtServerSecretKey("random");
+
+	            httpServletRequest.addHeader("Authorization", jwkstoken);
+
+	            authenticationFilter.doFilter(httpServletRequest, httpServletResponse, filterChain);
+	            assertEquals(HttpStatus.FORBIDDEN.value(),httpServletResponse.getStatus());
+	        }
+	        
 	  }
 }
